+++ .michele/IODELAY_Set.sh
@@ -1,24 +1,25 @@
 #!/bin/bash
 echo "Set all IODEAL to Default.. "  
 
-pci -w 9500 16
-pci -w 9504 f
-pci -w 9508 7
-pci -w 950C 0
-pci -w 9510 16
-pci -w 9514 f
-pci -w 9518 7
-pci -w 951C 0
-pci -w 9520 16
-pci -w 9524 f
-pci -w 9528 7
-pci -w 952C 0
-pci -w 9530 16
-pci -w 9534 f
-pci -w 9538 7
-pci -w 953C 0
+pci -w 9300 16
+pci -w 9304 f
+pci -w 9308 7
+pci -w 930C 0
+pci -w 9310 16
+pci -w 9314 f
+pci -w 9318 7
+pci -w 931C 0
+pci -w 9320 16
+pci -w 9324 f
+pci -w 9328 7
+pci -w 932C 0
+pci -w 9330 16
+pci -w 9334 f
+pci -w 9338 7
+pci -w 933C 0
 
-pci -w 9540 02
+pci -w 9340 02
+pci -w 9350 0000
 
 sleep 0.2
 pci -w 9040 80000002