+++ .michele/PWDown.sh
@@ -12,13 +12,4 @@ echo "--------------------------------"
 pci -w 9000 DF00
 pci -r 9010 -s 1
 
-sleep 0.2
-pci -w 9000 E000
-pci -r 9010 -s 1
-sleep 0.2
-
-pci -w 9000 e100
-pci -r 9010 -s 1
-sleep 0.2
-
 echo "DONE .. "