+++ .michele/Reset_all.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
-# By Michele Caselle for UFO 6 / 20 MPixels - camera
-echo "--------------------------------"
-echo "-------  	S T A R T  -----------"
-echo "-- C O N F I G U R A T I O N  --"
-echo "--------------------------------"
+# By Michele Caselle for UFO 6 / 20 MPixels - camera, date: August 2020
+	
+echo "--------------------------------------------------------------------"
+echo "--------------------------- S T A R T  -----------------------------"
+echo "-------------------- C O N F I G U R A T I O N  --------------------"
+echo "---------------------- ( UFO 6 - 20 MPixel ) -----------------------"
+echo "--------------------------- Normal Mode ----------------------------"
+echo "--------------------------------------------------------------------"
 
 pci -w 9020 100
 sleep 0.2
@@ -55,16 +58,7 @@ sleep 0.2
 pci -w 9000 FB66
 pci -r 9010 -s 1
 
-echo "Set the number of line to 3840 ... "  
-pci -w 9000 9a00
-sleep 0.2
-pci -r 9010 -s1
-sleep 0.2
-pci -w 9000 9b0F
-sleep 0.2
-pci -r 9010 -s1
-
-#echo "Set the start single of line to 255 ... "  
+echo "Set the start single of line to 0 ... "  
 #pci -w 9000 98FF
 pci -w 9000 9800
 sleep 0.2
@@ -74,12 +68,38 @@ pci -w 9000 9900
 sleep 0.2
 pci -r 9010 -s1
 
-echo "Set the exp time .. 4 ms"  
-pci -w 9000 a00f
+echo "Set the number of line to 3840 ... "  
+pci -w 9000 9a00
 sleep 0.2
 pci -r 9010 -s1
 sleep 0.2
-pci -w 9000 a100
+pci -w 9000 9b0F
+sleep 0.2
+pci -r 9010 -s1
+
+echo "Set the subsampling to 0 ... "  
+#pci -w 9000 98FF
+pci -w 9000 9C00
+sleep 0.2
+pci -r 9010 -s1
+sleep 0.2
+pci -w 9000 9D00
+sleep 0.2
+pci -r 9010 -s1
+pci -w 9000 9E00
+sleep 0.2
+pci -r 9010 -s1
+sleep 0.2
+pci -w 9000 9F00
+sleep 0.2
+pci -r 9010 -s1
+
+echo "Set the min. value of exp time ... 96 us"  
+pci -w 9000 a001
+sleep 0.2
+pci -r 9010 -s1
+sleep 0.2
+pci -w 9000 a1a0
 sleep 0.2
 pci -r 9010 -s1
 sleep 0.2
@@ -90,10 +110,7 @@ sleep 0.2
 pci -r 9010 -s1
 
 echo "FPGA Reset ... "	
-pci -w 9040 80000005
-sleep 0.2
-pci -w 9040 80000201
-sleep 0.2
+./reset_fpga.sh 
 
 echo "Set DMA timeout"
 pci -w 64 ffffff
@@ -104,7 +121,8 @@ pci -w 9000 9601
 sleep 0.2
 pci -r 9010 -s1
 
-echo "Set 1 frame for each frame requested"
-sleep 0.2
-
-echo "finish .. "  
+echo "--------------------------------------------------------------------"
+echo "finish .. "
+echo "--------------------------------------------------------------------"
+echo "status ..."  
+pci -r 9050 -s 8